+++ FO-DICOM.Core/DicomItemComparer.cs
@@ -3,7 +3,6 @@
 #nullable disable
 
 using FellowOakDicom.IO.Buffer;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -19,8 +18,15 @@ namespace FellowOakDicom
 
     public class DicomValueComparer : IEqualityComparer<DicomItem>
     {
+        public static DicomValueComparer DefaultInstance { get; set; } = new DicomValueComparer();
+
         public bool Equals(DicomItem item1, DicomItem item2)
         {
+            if (ReferenceEquals(item1, item2))
+            {
+                // short circuit comparing the same item
+                return true;
+            }
             if (item1 is DicomElement xElement && item2 is DicomElement yElement)
             {
                 if (xElement.Buffer is BulkDataUriByteBuffer xBulkbuffer && !xBulkbuffer.IsMemory || 
@@ -30,10 +36,7 @@ namespace FellowOakDicom
                     return item1.Tag == item2.Tag;
                 }
 
-                var xValue = string.Join("\\", xElement.Get<string[]>());
-                var yValue = string.Join("\\", yElement.Get<string[]>());
-
-                return item1.Tag == item2.Tag && xValue == yValue;
+                return xElement.Tag == yElement.Tag && xElement.Equals(yElement);
             }
 
             if (item1 is DicomSequence xSequence && item2 is DicomSequence ySequence)
@@ -44,12 +47,11 @@ namespace FellowOakDicom
                     return false;
                 }
 
-                var datasetComparer = new DicomDatasetComparer();
                 for (var i = 0; i < itemsCount; i++)
                 {
                     var dataset1 = xSequence.Items[i];
                     var dataset2 = ySequence.Items[i];
-                    if (!datasetComparer.Equals(dataset1, dataset2))
+                    if (!DicomDatasetComparer.DefaultInstance.Equals(dataset1, dataset2))
                     {
                         return false;
                     }
@@ -77,6 +79,9 @@ namespace FellowOakDicom
 
     public class DicomDatasetComparer : IEqualityComparer<DicomDataset>
     {
+
+        public static DicomDatasetComparer DefaultInstance { get; set; } = new DicomDatasetComparer();
+
         public bool Equals(DicomDataset dataset1, DicomDataset dataset2)
         {
             if ((dataset1 == null) != (dataset2 == null))
@@ -95,10 +100,10 @@ namespace FellowOakDicom
                 return false;
             }
 
-            var valueComparer = new DicomValueComparer();
-            foreach (var elements in dataset1.Zip(dataset2, Tuple.Create))
+            foreach (var element in dataset1)
             {
-                if (!valueComparer.Equals(elements.Item1, elements.Item2))
+                var element2 = dataset2.GetDicomItem<DicomItem>(element.Tag);
+                if (!DicomValueComparer.DefaultInstance.Equals(element, element2))
                 {
                     return false;
                 }